Refactoring to use commander instead of Vorpal
