+++ src/index.js
@@ -1,25 +1,46 @@
-import vorpal from 'vorpal';
+import yeoman from 'yeoman-environment';
+import program from 'commander';
 
-const commands = [
-  'generate'
-];
+const env = yeoman.createEnv();
 
-export default function() {
-  const app = vorpal();
+const feathersGenerators = 'generator-feathers/generators';
 
-  commands.map((cmd) => require(`./commands/${cmd}`).default(app));
+env.register(require.resolve(`${feathersGenerators}/app`), 'feathers:app');
+env.register(require.resolve(`${feathersGenerators}/authentication`), 'feathers:authentication');
+env.register(require.resolve(`${feathersGenerators}/connection`), 'feathers:connection');
+env.register(require.resolve(`${feathersGenerators}/hook`), 'feathers:hook');
+env.register(require.resolve(`${feathersGenerators}/middleware`), 'feathers:middleware');
+env.register(require.resolve(`${feathersGenerators}/service`), 'feathers:service');
+env.register(require.resolve('generator-feathers-plugin'), 'feathers:plugin');
 
-  if (process.argv.length > 2) {
-    // one and done
-    app.parse(process.argv);
+module.exports = function(argv, generatorOptions = {
+  disableNotifyUpdate: true
+}) {
+  program.version(require('../package.json').version)
+    .usage('generate [type]');
+
+  program.command('generate [type]')
+  .description(`Run a generator. Type can be
+  • app - Create a new Feathers application in the current folder
+  • authentication - Set up authentication for the current application
+  • connection - Initialize a new database connection
+  • hook - Create a new hook
+  • middleware - Create an Express middleware
+  • service - Generate a new service
+  • plugin - Create a new Feathers plugin
+`)
+    .action(type => {
+      if (!type) {
+        program.help();
+      } else {
+        env.run(`feathers:${type}`, generatorOptions);
+      }
+    });
+
+  program.command('*').action(() => program.help());
+  program.parse(argv);
+
+  if (argv.length === 2) {
+    program.help();
   }
-  else {
-    // interactive shell
-    app.log(`Welcome to the Feathers command line.`);
-    app.log('Type "exit" to quit, "help" for a list of commands.');
-
-    app
-      .delimiter('feathers$')
-      .show();
-  }
-}
+};
