--- test/mocha.opts
@@ -1,2 +1,0 @@
---recursive
---compilers js:babel-core/register
